+++ packages/components/src/components/link/link.tsx
@@ -8,10 +8,11 @@ import {
   throwIfInvalidLinkUsage,
   validateProps,
   LINK_BUTTON_VARIANTS,
+  hasVisibleIcon,
 } from '../../utils';
 import type {
   BreakpointCustomizable,
-  IconName,
+  LinkButtonIconName,
   LinkTarget,
   LinkVariant,
   PropTypes,
@@ -43,10 +44,10 @@ export class Link {
   @Element() public host!: HTMLElement;
 
   /** The style variant of the link. */
-  @Prop() public variant?: LinkVariant = 'secondary';
+  @Prop() public variant?: LinkVariant = 'primary';
 
-  /** The icon shown. */
-  @Prop() public icon?: IconName = 'arrow-head-right';
+  /** The icon shown. By choosing 'none', no icon is displayed. */
+  @Prop() public icon?: LinkButtonIconName = 'none';
 
   /** A URL path to a custom icon. */
   @Prop() public iconSource?: string;
@@ -78,7 +79,7 @@ export class Link {
 
   public render(): JSX.Element {
     validateProps(this, propTypes);
-    attachComponentCss(this.host, getComponentCss, this.variant, this.hideLabel, !this.href, this.theme);
+    attachComponentCss(this.host, getComponentCss, this.icon, this.variant, this.hideLabel, !this.href, this.theme);
 
     const TagType = this.href === undefined ? 'span' : 'a';
     const PrefixedTagNames = getPrefixedTagNames(this.host);
@@ -94,21 +95,23 @@ export class Link {
           ...parseAndGetAriaAttributes(this.aria),
         })}
       >
-        <PrefixedTagNames.pIcon
-          class="icon"
-          size="inherit"
-          name={this.icon}
-          source={this.iconSource}
-          color="inherit"
-          theme={
-            this.variant === 'tertiary'
-              ? this.theme
-              : this.variant === 'secondary' && this.theme === 'dark'
-              ? 'light'
-              : 'dark'
-          } // relevant for ssr support
-          aria-hidden="true"
-        />
+        {(hasVisibleIcon(this.icon) || this.iconSource || this.hideLabel) &&
+          <PrefixedTagNames.pIcon
+            class="icon"
+            size="inherit"
+            name={hasVisibleIcon(this.icon) ? this.icon : 'arrow-right'}
+            source={this.iconSource}
+            color="inherit"
+            theme={
+              this.variant === 'tertiary'
+                ? this.theme
+                : this.variant === 'secondary' && this.theme === 'dark'
+                ? 'light'
+                : 'dark'
+            } // relevant for ssr support
+            aria-hidden="true"
+          />
+        }
         <span>
           <slot />
         </span>