+++ packages/components/src/styles/host-hidden-styles.spec.ts
@@ -1,0 +1,7 @@
+import { hostHiddenStyles } from './host-hidden-styles';
+
+describe('hostHiddenStyles', () => {
+  it('should return correct jss styles', () => {
+    expect(hostHiddenStyles).toMatchSnapshot();
+  });
+});