+++ packages/components/CHANGELOG.md
@@ -15,6 +15,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Changed
 
+- Default value of prop `variant` of `Button` and `Link` has changed from "secondary" to "primary"
+- Default value of prop `icon` of `Button` and `Link` has changed from "arrow-head-right" to "none"
 - `Spinner`, `Link Pure`, `Button Pure` matches new design language
 - Prop `weight` for `Link Pure` and `Button Pure` is deprecated now, only regular font weight will be applied